+++ docs/reference/changelog.rst
@@ -49,6 +49,8 @@ Services
   ``ProjectService.project_vars`` property instead.
 - Registering services on ``ServiceFactory`` instantiation is removed. Use
   ``ServiceFactory.register()`` instead.
+- ``ServiceFactory.set_kwargs()`` is removed. Use ``ServiceFactory.update_kwargs()``
+  instead.
 
 For a complete list of commits, check out the `7.0.0`_ release on GitHub.